feat(pending-outflows): create PrintCheckDialog from PrintCheckButton
Rename PrintCheckButton.tsx to PrintCheckDialog.tsx. Move settings
and open state to props: settings, expense, onOpenChange.

Add isOpen, wasOpen, and shownExpense. Reset the form at render time
on each open and on each row switch. Keep the account-init effect as
a late fill for defaultAccount.

Read displayExpense in the body. Return null only when displayExpense
is null, so the dialog exit animation shows the last expense.

Delete the trigger button. PendingOutflowCard will render its own
button and call onPrintCheck in a later step.

Keep handlePrint in its exact order: fetch MICR secrets, claim the
check number, update the outflow, write the audit row, build and
save the PDF.

PendingOutflowCard.tsx still imports the deleted PrintCheckButton.
The next step wires PendingOutflowsList and PendingOutflowCard to
this new component and fixes that import.
