--- src/components/pending-outflows/PrintCheckButton.tsx
@@ -1,290 +1,0 @@
-import { useState, useEffect } from 'react';
-
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-} from '@/components/ui/dialog';
-import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select';
-
-import { Printer, FileText, Loader2 } from 'lucide-react';
-
-import { useRestaurantContext } from '@/contexts/RestaurantContext';
-import { useCheckSettings } from '@/hooks/useCheckSettings';
-import { useCheckBankAccounts } from '@/hooks/useCheckBankAccounts';
-import { useCheckAuditLog } from '@/hooks/useCheckAuditLog';
-import { usePendingOutflowMutations } from '@/hooks/usePendingOutflows';
-import { SearchableAccountSelector } from '@/components/banking/SearchableAccountSelector';
-import {
-  generateCheckPDF,
-  generateCheckPDFAsync,
-  generateCheckFilename,
-  buildPrintConfig,
-  numberToWords,
-} from '@/utils/checkPrinting';
-import { formatCurrency } from '@/utils/pdfExport';
-import { toast } from 'sonner';
-
-import type { PendingOutflow } from '@/types/pending-outflows';
-
-interface PrintCheckButtonProps {
-  expense: PendingOutflow;
-}
-
-export function PrintCheckButton({ expense }: PrintCheckButtonProps) {
-  const { selectedRestaurant } = useRestaurantContext();
-  const { settings } = useCheckSettings();
-  const { accounts, defaultAccount, claimCheckNumbers: claimForAccount, fetchAccountSecrets } = useCheckBankAccounts();
-  const { logCheckAction } = useCheckAuditLog();
-  const { updatePendingOutflow } = usePendingOutflowMutations();
-
-  const [open, setOpen] = useState(false);
-  const [memo, setMemo] = useState(expense.notes ?? '');
-  const [isPrinting, setIsPrinting] = useState(false);
-  const [selectedAccountId, setSelectedAccountId] = useState<string | null>(null);
-  const [selectedCategoryId, setSelectedCategoryId] = useState<string | null>(
-    expense.category_id ?? null,
-  );
-
-  // Reset dialog-local fields each time the dialog opens (or expense changes while open)
-  useEffect(() => {
-    if (!open) return;
-    setMemo(expense.notes ?? '');
-    setSelectedCategoryId(expense.category_id ?? null);
-  }, [open, expense.notes, expense.category_id]);
-
-  // Initialize account selection once per open cycle
-  useEffect(() => {
-    if (!open) return;
-    setSelectedAccountId((current) => current ?? defaultAccount?.id ?? null);
-  }, [open, defaultAccount?.id]);
-
-  // Don't show if settings aren't configured
-  if (!settings) return null;
-
-  const handlePrint = async () => {
-    if (!settings || !selectedRestaurant) return;
-
-    const selectedAccount = accounts.find((a) => a.id === selectedAccountId) ?? defaultAccount;
-    if (!selectedAccount) {
-      toast.error('Please select a bank account');
-      return;
-    }
-
-    setIsPrinting(true);
-    try {
-      // Fetch MICR secrets first so any failure aborts BEFORE we claim a check
-      // number or write a "printed" audit row that wouldn't match a real PDF.
-      let secrets: { routing_number: string; account_number: string } | null = null;
-      if (selectedAccount.print_bank_info) {
-        if (!selectedAccount.routing_number || !selectedAccount.account_number_last4) {
-          toast.error('Bank info incomplete. Open Check Settings to add the routing and account numbers, or turn off "Print bank info" for this account.');
-          return;
-        }
-        try {
-          secrets = await fetchAccountSecrets(selectedAccount.id);
-        } catch (e) {
-          toast.error(e instanceof Error ? e.message : "Couldn't load bank info");
-          return;
-        }
-        if (!secrets) {
-          toast.error('Account number is missing. Re-enter it in Check Settings.');
-          return;
-        }
-      }
-
-      const checkNumber = await claimForAccount.mutateAsync({
-        accountId: selectedAccount.id,
-        count: 1,
-      });
-
-      await updatePendingOutflow.mutateAsync({
-        id: expense.id,
-        input: {
-          payment_method: 'check',
-          reference_number: String(checkNumber),
-          notes: memo.trim() || expense.notes,
-          check_bank_account_id: selectedAccount.id,
-          category_id: selectedCategoryId,
-        },
-      });
-
-      await logCheckAction.mutateAsync({
-        check_number: checkNumber,
-        payee_name: expense.vendor_name,
-        amount: expense.amount,
-        issue_date: expense.issue_date,
-        memo: memo.trim() || null,
-        action: 'printed',
-        pending_outflow_id: expense.id,
-        check_bank_account_id: selectedAccount.id,
-      });
-
-      const checks = [
-        {
-          checkNumber,
-          payeeName: expense.vendor_name,
-          amount: expense.amount,
-          issueDate: expense.issue_date,
-          memo: memo.trim() || undefined,
-        },
-      ];
-      const config = buildPrintConfig(settings, selectedAccount, secrets);
-      const pdf = selectedAccount.print_bank_info
-        ? await generateCheckPDFAsync(config, checks)
-        : generateCheckPDF(config, checks);
-      const filename = generateCheckFilename(selectedRestaurant.restaurant.name, [checkNumber]);
-      pdf.save(filename);
-
-      toast.success(`Check #${checkNumber} printed`);
-      setOpen(false);
-    } catch (err) {
-      console.error('Print check error:', err);
-      toast.error(err instanceof Error ? err.message : 'Failed to print check');
-    } finally {
-      setIsPrinting(false);
-    }
-  };
-
-  return (
-    <>
-      <Button
-        variant="outline"
-        size="sm"
-        onClick={(e) => {
-          e.stopPropagation();
-          setOpen(true);
-        }}
-        className="h-8 px-2 rounded-lg text-[13px] font-medium"
-        aria-label={`Print check for ${expense.vendor_name}`}
-      >
-        <Printer className="h-3.5 w-3.5 mr-1" />
-        Print
-      </Button>
-
-      <Dialog open={open} onOpenChange={setOpen}>
-        <DialogContent className="max-w-md max-h-[80vh] overflow-y-auto p-0 gap-0 border-border/40">
-          <DialogHeader className="px-6 pt-6 pb-4 border-b border-border/40">
-            <div className="flex items-center gap-3">
-              <div className="h-10 w-10 rounded-xl bg-muted/50 flex items-center justify-center">
-                <FileText className="h-5 w-5 text-foreground" />
-              </div>
-              <div>
-                <DialogTitle className="text-[17px] font-semibold text-foreground">
-                  Print Check
-                </DialogTitle>
-                <DialogDescription className="text-[13px] text-muted-foreground mt-0.5">
-                  Review details before printing
-                </DialogDescription>
-              </div>
-            </div>
-          </DialogHeader>
-
-          <div className="px-6 py-5 space-y-5">
-            {/* Summary */}
-            <div className="rounded-xl border border-border/40 bg-muted/30 p-4 space-y-3">
-              <div className="flex items-center justify-between">
-                <span className="text-[13px] text-muted-foreground">Pay to</span>
-                <span className="text-[14px] font-medium text-foreground">{expense.vendor_name}</span>
-              </div>
-              <div className="flex items-center justify-between">
-                <span className="text-[13px] text-muted-foreground">Amount</span>
-                <div className="text-right">
-                  <p className="text-[17px] font-semibold text-foreground">
-                    {formatCurrency(expense.amount)}
-                  </p>
-                  <p className="text-[11px] text-muted-foreground">{numberToWords(expense.amount)}</p>
-                </div>
-              </div>
-            </div>
-
-            {/* Bank account selector — only shown when multiple accounts exist */}
-            {accounts.length > 1 && (
-              <div className="space-y-2">
-                <Label htmlFor="print-check-account" className="text-[12px] font-medium text-muted-foreground uppercase tracking-wider">
-                  Bank Account
-                </Label>
-                <Select value={selectedAccountId ?? ''} onValueChange={setSelectedAccountId}>
-                  <SelectTrigger id="print-check-account" className="h-10 text-[14px] bg-muted/30 border-border/40 rounded-lg">
-                    <SelectValue placeholder="Select account" />
-                  </SelectTrigger>
-                  <SelectContent>
-                    {accounts.map((account) => (
-                      <SelectItem key={account.id} value={account.id}>
-                        {account.account_name}{account.bank_name ? ` (${account.bank_name})` : ''}
-                      </SelectItem>
-                    ))}
-                  </SelectContent>
-                </Select>
-              </div>
-            )}
-
-            {/* Memo */}
-            <div className="space-y-2">
-              <Label htmlFor="print-check-memo" className="text-[12px] font-medium text-muted-foreground uppercase tracking-wider">
-                Memo (optional)
-              </Label>
-              <Input
-                id="print-check-memo"
-                value={memo}
-                onChange={(e) => setMemo(e.target.value)}
-                placeholder="Purpose of payment"
-                className="h-10 text-[14px] bg-muted/30 border-border/40 rounded-lg focus-visible:ring-1 focus-visible:ring-border"
-              />
-            </div>
-
-            {/* Category */}
-            <div className="space-y-2">
-              <Label htmlFor="print-check-category" className="text-[12px] font-medium text-muted-foreground uppercase tracking-wider">
-                Category (optional)
-              </Label>
-              <SearchableAccountSelector
-                triggerId="print-check-category"
-                value={selectedCategoryId ?? undefined}
-                onValueChange={(v) => setSelectedCategoryId(v || null)}
-                filterByTypes={['expense', 'cogs', 'asset']}
-                placeholder="Pick a chart-of-accounts category"
-                triggerAriaLabel="Category for this check"
-              />
-            </div>
-          </div>
-
-          {/* Footer */}
-          <div className="px-6 py-4 border-t border-border/40 flex justify-end gap-2">
-            <Button
-              variant="outline"
-              onClick={() => setOpen(false)}
-              disabled={isPrinting}
-              className="h-9 px-4 rounded-lg text-[13px] font-medium text-muted-foreground hover:text-foreground"
-            >
-              Cancel
-            </Button>
-            <Button
-              onClick={handlePrint}
-              disabled={isPrinting}
-              className="h-9 px-4 rounded-lg bg-foreground text-background hover:bg-foreground/90 text-[13px] font-medium"
-            >
-              {isPrinting ? (
-                <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-              ) : (
-                <Printer className="mr-2 h-4 w-4" />
-              )}
-              Print Check
-            </Button>
-          </div>
-        </DialogContent>
-      </Dialog>
-    </>
-  );
-}
